requirements.txt
APScheduler==3.2.0
Flask==0.11.1
requests==2.11.0
lxml==4.1.0

pymongo==3.2.2
redis==2.10.5


